Stopped being a dunce (I think)
In my defense, I forgot why delegates work so well.
